First running version.
This commit includes a running version of the application that automaticallygenerates based on the number of tiles provided for a given area
the <div> for the maps, so the user can pick one.

The script renders everything dynamically, so we can actually vary the number of tiles without problems. The oly limitation could be the number of tiles to show in one unique place.